component files renamed, convert to persian number fixed
