+++ src/components/Leaderboard/Leaderboard.tsx
@@ -1,0 +1,33 @@
+import { styled, Box } from '@mui/material';
+import TopRanks from './TopRanks/TopRanks'
+import OtherRanks from './OtherRanks/OtherRanks';
+import { normalizeData } from '../../utils/normalizeData'
+import background from '../../assets/background.svg';
+import { Data } from '../../utils/types'
+import { rawData1, rawData2, rawData3, rawData4, rawData5, rawData6, rawData7 } from '../../data';
+
+const Root = styled(Box)({
+  width: '100%',
+  height: '100vh',
+  backgroundImage: `url(${background})`,
+  backgroundRepeat: 'no-repeat',
+  backgroundPositionY: 10,
+  backgroundSize: 'contain',
+  position: 'relative',
+});
+
+const Leaderboard = () => {
+  const rawData: Data = [...rawData1]
+  const { firstRanks, secondRanks, thirdRanks, otherRanks } = normalizeData(rawData)
+
+  return (
+    <Root>
+      <TopRanks data={firstRanks} size={3} componentDirection={'row'} />
+      {(secondRanks.length > 0)  && (<TopRanks data={secondRanks} size={2.5} componentDirection={'row-reverse'} />)}
+      {(thirdRanks.length > 0)  && (<TopRanks data={thirdRanks} size={2} componentDirection={'row'} />)}
+      {(otherRanks.length > 0) && (<OtherRanks data={otherRanks} />)}
+    </Root>
+  );
+};
+
+export default Leaderboard;
